+++ src/main/java/com/example/hrisapi/dto/response/ReportTagihanGajiResponse.java
@@ -21,7 +21,7 @@ public class ReportTagihanGajiResponse {
 
     private String namaProyek;
 
-    private Double gaji;
+    private Long gaji;
 
     private Double tunjangan;
